--- src/main/java/com/dnd/bbok/basicchecklist/application/port/in/GetBasicChecklistQuery.java
@@ -1,5 +1,0 @@
-package com.dnd.bbok.basicchecklist.application.port.in;
-
-public interface GetBasicChecklistQuery {
-    GetBasicChecklistResponse getBasicChecklist();
-}
