+++ src/main/java/com/dnd/bbok/basicchecklist/application/port/in/usecase/GetBasicChecklistQuery.java
@@ -1,0 +1,7 @@
+package com.dnd.bbok.basicchecklist.application.port.in.usecase;
+
+import com.dnd.bbok.basicchecklist.application.port.in.response.GetBasicChecklistResponse;
+
+public interface GetBasicChecklistQuery {
+    GetBasicChecklistResponse getBasicChecklist();
+}
