--- src/main/java/com/dnd/bbok/member/application/port/in/GetMemberChecklistQuery.java
@@ -1,7 +1,0 @@
-package com.dnd.bbok.member.application.port.in;
-
-import java.util.UUID;
-
-public interface GetMemberChecklistQuery {
-    GetMemberChecklistResponse getMemberChecklistQuery(UUID memberId);
-}
