--- src/main/java/com/dnd/bbok/member/application/port/in/GetMemberQuery.java
@@ -1,9 +1,0 @@
-package com.dnd.bbok.member.application.port.in;
-
-import java.util.UUID;
-
-public interface GetMemberQuery {
-
-  GetMemberInfoResponse getMember(UUID memberId);
-
-}
