+++ src/main/java/com/dnd/bbok/member/application/port/in/usecase/GetMemberChecklistQuery.java
@@ -1,0 +1,9 @@
+package com.dnd.bbok.member.application.port.in.usecase;
+
+import com.dnd.bbok.member.application.port.in.response.GetMemberChecklistResponse;
+
+import java.util.UUID;
+
+public interface GetMemberChecklistQuery {
+    GetMemberChecklistResponse getMemberChecklistQuery(UUID memberId);
+}
