+++ src/main/java/com/dnd/bbok/member/application/service/MemberTestService.java
@@ -1,6 +1,11 @@
 package com.dnd.bbok.member.application.service;
 
-import com.dnd.bbok.member.application.port.in.*;
+import com.dnd.bbok.member.application.port.in.request.CreateMemberChecklistRequest;
+import com.dnd.bbok.member.application.port.in.response.GetMemberChecklistResponse;
+import com.dnd.bbok.member.application.port.in.response.GetMemberInfoResponse;
+import com.dnd.bbok.member.application.port.in.usecase.CreateMemberChecklistUseCase;
+import com.dnd.bbok.member.application.port.in.usecase.GetMemberChecklistQuery;
+import com.dnd.bbok.member.application.port.in.usecase.GetMemberQuery;
 import com.dnd.bbok.member.application.port.out.LoadMemberChecklistPort;
 import com.dnd.bbok.member.application.port.out.LoadMemberPort;
 import com.dnd.bbok.member.application.port.out.SaveMemberChecklistPort;
